soc: qcom: pil-q6v5-mss: Do not free memory allocated to modem pil
Modem now uses new memory feature(fixup-reserve-region). Instead of
fixed carve-out, region can adapt and shrink to actual memory
requirement after first allocation. This memory is not considered
as carve out and is exposed to kernel. When memory is freed, kernel
maps the region as normal cacheable memory. This results in XPU
violation errors because the region is still marked as XPU protected
by MBA. Hence modem pil is avoiding freeing of memory and reusing
the allocated region.

Change-Id: I507f4a09e6f86dc8584fb23877fe9eab51ef9bb6
Signed-off-by: Arun KS <arunks@codeaurora.org>