arm: DT: msm8939-gpu: Various fixes to boot for devbw and kgsl
Add active-only flag to bwmon, so that we can release the bus vote
when the CPU subsystem is inactive (and not race).
Add a timeout to enter deeper power saving state from NAP to allow
the device to sleep correctly.
(QUIRK) Add empty ocmem-bus-client so that kernel will be happy.