media: msm: camera_v2: Do not drop reconfigure requests 630/660
ISP reconfigure requests may come while the MCT is still
processing some frame: in this case Qualcomm tells us to
drop the request until the MCT processing is done.

That ain't working fine at all on (old only?) platforms
that are using a Dual-VFE configuration, as the ISP seems
to slow down a whole lot during error processing, so the
MCT will be *always* signaling that's processing frames,
never actually giving a chance to this driver to actually
reconfigure and/or recover from the bad hardware state.
For this reason, just never drop reconfigure requests,
process them and do the job in any case: there are no
measurable drawbacks to this and the user experience is
anyway even improved.

While another cleaner solution to this issue does indeed
exist, it is way more complicated than whatever gets done
here.. and it doesn't give *any* improvement compared to
this patch anyway, so let's keep it simple and short.