arm64: dts: somc: Add device tree support for Xperia 10+ (Mermaid)
This is an initial bringup for Ganges platform Mermaid device.
The reason the Xperia 10 (Kirin) is missing here is that it is based on
the SDM630 SoC and the dts are not yet pushed here. It will be added
after the initialization of the dts files for the SDM630 SoC.

This snapshot is based on the implementation from the
SODP 4.14 kernel.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>